Shibuya City, Tokyo, Japan
I took this at the Meiji Jingu Shinto shrine in Shibuya.  I saw this type of elevated threshold a lot when I was in China.  Two things that I was told were that it is disrespectful to step on it and that it is there to keep out evil spirits.