Paju-si, Gyeonggi-do, South Korea
This is the "Bridge of Freedom." It crosses a lake at Imjingak Pyeonghwa Nuri Park. Initially, the Bridge had two spans that headed up and down, but they were destroyed by bombing, and only the pier remained. Later, the west Bridge was restored to exchange prisoners of war, and a temporary bridge was installed at the south end. 13,000 POWs were exchanged on it.