Paju-si, Gyeonggi-do, South Korea
I took this at the New Dora Observatory, along the DMZ in South Korea. The original Dora Observatory, which this one replaces, was just a few yards from the bus. They asked us not to photograph it because it is part of an active military installation.  The line you can see going through the image is the South Korean side of the DMZ.