Venice, Italy
I took this at Saint Mark's Basilica (Basilica di San Marco) in Venice. Originally, this building was intended as an extension of the Doge’s Palace; however, the construction of the Basilica of St. Mark, which began in 828 and concluded in 832, was undertaken to house the body of St. Mark the Apostle, who was brought from Alexandria and named protector of the city.