Kyoto, Japan
I captured this at Kiyomizudera Temple in Kyoto. It was established in 780 on the site of the Otowa Waterfall in the wooded hills east of Kyoto and takes its name from the waterfall's waters. The temple was originally associated with the Hosso sect, one of the oldest schools of Japanese Buddhism, but created its own Kita Hosso sect in 1965.